requirements.txt
elastic-transport==8.13.1
elasticsearch==8.14.0
streamlit==1.37.
youtube-transcript-api==0.4.1
python-dotenv==0.19.1
aiohttp==3.7.4
eland[pytorch]==8.15.0
google-api-python-client
# jello==1.14.1
# openai==1.35.13
# pprint==0.7.0
cohere==4.1.0